requirements.txt
django
python-decouple
django-taggit
psycopg2-binary
dj-database_url
django-browser-reload
markdown